requirements.txt
Flask
flask-sqlalchemy
flask-login
werkzeug
gunicorn
python-dotenv
psycopg2-binary
reportlab
Pillow